+++ wire-ios-request-strategy/Sources/Payloads/Conversation/Payload+Conversation.swift
@@ -214,29 +214,6 @@ extension Payload {
         let data: String
     }
 
-    struct UpdateConversationMLSMessageAdd: Codable {
-
-        enum CodingKeys: String, CodingKey {
-            case id = "conversation"
-            case qualifiedID = "qualified_conversation"
-            case from
-            case qualifiedFrom = "qualified_from"
-            case subconversationType = "subconv"
-            case timestamp = "time"
-            case type
-            case data
-        }
-
-        let id: UUID
-        let qualifiedID: QualifiedID?
-        let from: UUID
-        let qualifiedFrom: QualifiedID?
-        let subconversationType: SubgroupType?
-        let timestamp: Date
-        let type: String
-        let data: String
-    }
-
     struct UpdateConversationProtocolChange: CodableEventData {
         static var eventType: ZMUpdateEventType {
             return .conversationConnectRequest
